+++ packages/lex/lex-schema/src/util/memoize.ts
@@ -1,57 +1,37 @@
 /*@__NO_SIDE_EFFECTS__*/
-export function memoizedOptions<
-  F extends (options?: NonNullable<unknown>) => any,
->(
-  fn: F,
-  keyFn?: (
-    options: NonNullable<Parameters<F>[0]>,
-  ) => string | number | boolean | null | void,
-): F {
-  let emptyOptionsValue: ReturnType<F> | undefined
+export function memoizedOptions<F extends (...args: any[]) => any>(fn: F): F {
+  let cache: null | { value: ReturnType<F> } = null
 
-  if (keyFn) {
-    const cache = new Map<string | number | boolean | null, ReturnType<F>>()
-    const fromCache = (
-      options: NonNullable<Parameters<F>[0]>,
-    ): ReturnType<F> => {
-      const key = keyFn(options)
-      if (key !== undefined) {
-        const cached = cache.get(key)
-        if (cached) return cached
-        const result = fn(options) as ReturnType<F>
-        cache.set(key, result)
-        return result
-      }
-
-      return fn(options)
+  return function cached(...args: any[]): ReturnType<F> {
+    // Not using the cache if there are args
+    if (args.length > 0) {
+      return fn(...args)
     }
 
-    return ((options: Parameters<F>[0]): ReturnType<F> => {
-      // Non-empty options case
-      if (options) for (const _ in options) return fromCache(options)
+    if (cache != null) {
+      return cache.value
+    }
 
-      // Empty or missing options case
-      return (emptyOptionsValue ??= fn(options))
-    }) as F
-  }
-
-  return ((options: Parameters<F>[0]): ReturnType<F> => {
-    // Non-empty options case
-    if (options) for (const _ in options) return fn(options)
-
-    // Empty or missing options case
-    return (emptyOptionsValue ??= fn(options))
-  }) as F
+    const value = fn(...args)
+    cache = { value }
+    return value
+  } as F
 }
 
 /*@__NO_SIDE_EFFECTS__*/
-export function memoizedTransformer<F extends (arg: any) => any>(fn: F): F {
-  const cache = new WeakMap<object, ReturnType<F>>()
-  return ((arg: Parameters<F>[0]): ReturnType<F> => {
-    const cached = cache.get(arg)
+export function memoizedTransformer<
+  F extends (key: any, ...args: any[]) => unknown,
+>(fn: F): F {
+  let cache: WeakMap<object, ReturnType<F>>
+
+  return function cached(key: any, ...args: any[]): any {
+    if (args.length > 0) return fn(key, ...args)
+
+    cache ??= new WeakMap()
+    const cached = cache.get(key)
     if (cached) return cached
-    const result = fn(arg) as ReturnType<F>
-    cache.set(arg, result)
+    const result = fn(key, ...args) as ReturnType<F>
+    cache.set(key, result)
     return result
-  }) as F
+  } as F
 }