+++ packages/lex/lex-resolver/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @atproto/lex-resolver
 
+## 0.0.10
+
+### Patch Changes
+
+- Updated dependencies [[`dfd4bee`](https://github.com/bluesky-social/atproto/commit/dfd4bee4abbc1a3e53531bb499a6f3169c13ed9e), [`dfd4bee`](https://github.com/bluesky-social/atproto/commit/dfd4bee4abbc1a3e53531bb499a6f3169c13ed9e), [`dfd4bee`](https://github.com/bluesky-social/atproto/commit/dfd4bee4abbc1a3e53531bb499a6f3169c13ed9e), [`dfd4bee`](https://github.com/bluesky-social/atproto/commit/dfd4bee4abbc1a3e53531bb499a6f3169c13ed9e)]:
+  - @atproto/lex-data@0.0.8
+  - @atproto/lex-client@0.0.9
+  - @atproto/lex-document@0.0.10
+  - @atproto/lex-schema@0.0.9
+
 ## 0.0.9
 
 ### Patch Changes